% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{DataOptions}
\alias{DataOptions}
\alias{DataOptions<-}
\alias{DataOptions,MOFAmodel-method}
\alias{DataOptions<-,MOFAmodel,list-method}
\title{DataOptions: set and retrieve data options}
\usage{
DataOptions(object)

DataOptions(object) <- value

\S4method{DataOptions}{MOFAmodel}(object)

\S4method{DataOptions}{MOFAmodel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{a list with data options}
}
\value{
list of data options
}
\description{
Function to set and retrieve data options.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
DataOptions(MOFAobject)
}
