% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{Dimensions}
\alias{Dimensions}
\alias{Dimensions<-}
\alias{Dimensions,MOFAmodel-method}
\alias{Dimensions<-,MOFAmodel,list-method}
\title{Dimensions: set and retrieve dimensions}
\usage{
Dimensions(object)

Dimensions(object) <- value

\S4method{Dimensions}{MOFAmodel}(object)

\S4method{Dimensions}{MOFAmodel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{list of dimensions}
}
\value{
list of dimensions of the model
}
\description{
Function to set and retrieve model dimensions.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
Dimensions(MOFAobject)
}
