% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{InputData}
\alias{InputData}
\alias{InputData<-}
\alias{InputData,MOFAmodel-method}
\alias{InputData<-,MOFAmodel,MultiAssayExperiment-method}
\title{InputData: set and retrieve input data}
\usage{
InputData(object)

InputData(object) <- value

\S4method{InputData}{MOFAmodel}(object)

\S4method{InputData}{MOFAmodel,MultiAssayExperiment}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{a multi-assay experiment}
}
\value{
input data as a multi-assay experiment
}
\description{
Function to set and retrieve input data.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
InputData(MOFAobject)
}
