% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{Status}
\alias{Status}
\alias{Status<-}
\alias{Status,MOFAmodel-method}
\alias{Status<-,MOFAmodel,character-method}
\title{Status: set and retrieve training status}
\usage{
Status(object)

Status(object) <- value

\S4method{Status}{MOFAmodel}(object)

\S4method{Status}{MOFAmodel,character}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{character indicating whether the object is trained or untrained}
}
\value{
character indicating whether the object is trained or untrained
}
\description{
Function to set and retrieve taining status.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
Status(MOFAobject)
}
