% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{TrainData}
\alias{TrainData}
\alias{TrainData<-}
\alias{TrainData,MOFAmodel-method}
\alias{TrainData<-,MOFAmodel,list-method}
\title{TrainData: set and retrieve training data}
\usage{
TrainData(object)

TrainData(object) <- value

\S4method{TrainData}{MOFAmodel}(object)

\S4method{TrainData}{MOFAmodel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{a list of matrices containing the training data}
}
\value{
list of matrices containing the training data
}
\description{
Function to set and retrieve taining data.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
str(TrainData(MOFAobject))
}
