% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{TrainOptions}
\alias{TrainOptions}
\alias{TrainOptions<-}
\alias{TrainOptions,MOFAmodel-method}
\alias{TrainOptions<-,MOFAmodel,list-method}
\title{TrainOptions: set and retrieve training options}
\usage{
TrainOptions(object)

TrainOptions(object) <- value

\S4method{TrainOptions}{MOFAmodel}(object)

\S4method{TrainOptions}{MOFAmodel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{a list with training options}
}
\value{
list of training options
}
\description{
Function to set and retrieve training options.
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
TrainOptions(MOFAobject)
}
