% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\docType{methods}
\name{TrainStats}
\alias{TrainStats}
\alias{TrainStats<-}
\alias{TrainStats,MOFAmodel-method}
\alias{TrainStats<-,MOFAmodel,list-method}
\title{TrainStats: set and retrieve training statistcs}
\usage{
TrainStats(object)

TrainStats(object) <- value

\S4method{TrainStats}{MOFAmodel}(object)

\S4method{TrainStats}{MOFAmodel,list}(object) <- value
}
\arguments{
\item{object}{a \code{\link{MOFAmodel}} object.}

\item{value}{list of different training statistics (EBLO, number of factors)}
}
\value{
list of different training statistics (EBLO, number of factors)
}
\description{
Function to set and retrieve taining statistics
}
\examples{
# load a trained MOFAmodel object
filepath <- system.file("extdata", "scMT_model.hdf5", package = "MOFAdata")
MOFAobject <- loadModel(filepath)
str(TrainStats(MOFAobject))
}
