% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSEA.R
\name{plotEnrichment}
\alias{plotEnrichment}
\title{Line plot of Feature Set Enrichment Analysis results}
\usage{
plotEnrichment(object, fsea.results, factor, alpha = 0.1,
  max.pathways = 25, adjust = TRUE)
}
\arguments{
\item{object}{\code{\link{MOFAmodel}} object on which the Feature Set Enrichment Analyisis was performed}

\item{fsea.results}{output of \link{runEnrichmentAnalysis} function}

\item{factor}{Factor}

\item{alpha}{p.value threshold to filter out feature sets}

\item{max.pathways}{maximum number of enriched pathways to display}

\item{adjust}{use adjusted p-values?}
}
\value{
a \code{ggplot2} object
}
\description{
Method to plot Feature Set Enrichment Analyisis results for specific factors
}
\examples{
# Example on the CLL data
filepath <- system.file("extdata", "CLL_model.hdf5", package = "MOFAdata")
MOFA_CLL <- loadModel(filepath)

# perform Enrichment Analysis on mRNA data using pre-build Reactome gene sets
data("reactomeGS", package = "MOFAdata")
fsea.results <- runEnrichmentAnalysis(MOFA_CLL, view="mRNA", feature.sets=reactomeGS)

# Plot top 10 enriched pathwyas on factor 5:
plotEnrichment(MOFA_CLL, fsea.results, factor=5,  max.pathways=10)
}
